requirements.txt
certifi==2018.10.15
chardet==3.0.4
coverage==4.5.1
Django==2.1.2
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==2.4.0
django-filter==2.0.0
django-paypal==0.5.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-simple-history==2.5.1
djangorestframework==3.8.2
djangorestframework-jsonapi==2.6.0
PyJWT==1.5.2
idna==2.7
inflection==0.3.1
psycopg2==2.7.5
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.24
djangorestframework-simplejwt==3.3
paypalrestsdk==1.13.1
coverage==4.5.1
django-background-tasks==1.2.0
django-simple-history==2.5.1